refactor: consolidate account API client
