+++ app/routes/_index.tsx
@@ -1,7 +1,7 @@
 import type { LoaderFunctionArgs } from "react-router";
 import { Link, useLoaderData } from "react-router";
 
-import { account } from "~/lib/account";
+import { account } from "~/lib/api/account";
 import { getToken } from "~/lib/api/getToken";
 import { parseToken } from "~/lib/parseToken";
