+++ app/routes/accounts.$id._index.tsx
@@ -6,8 +6,8 @@ import { FollowButton } from "~/components/followButton";
 import { LoadMoreNoteButton } from "~/components/loadMoreNote";
 import type { NoteProps } from "~/components/note";
 import { Note } from "~/components/note";
-import type { AccountResponse } from "~/lib/account";
-import { account, accountTimeline } from "~/lib/account";
+import type { AccountResponse } from "~/lib/api/account";
+import { account, accountTimeline } from "~/lib/api/account";
 import { getToken } from "~/lib/api/getToken";
 import { loggedInAccount } from "~/lib/api/loggedInAccount";
 import { fetchNote } from "~/lib/api/note";
