+++ app/routes/accounts.$id.following.tsx
@@ -1,7 +1,7 @@
 import { useLoaderData, type LoaderFunctionArgs } from "react-router";
 
 import { FollowAccount } from "~/components/followAccount";
-import { account } from "~/lib/account";
+import { account } from "~/lib/api/account";
 import { getFollowingList, type FollowingResponse } from "~/lib/api/followlist";
 import { getToken } from "~/lib/api/getToken";
